[Cleanup] Restructure Source Files
