+++ SwiftFlowGitHubBrowser/Routes/Routes.swift
@@ -196,25 +196,3 @@ class MainViewRoutable: Routable {
 class RepositoryDetailRoutable: Routable {}
 class BookmarkRoutable: Routable {}
 class OAuthRoutable: Routable {}
-
-extension UINavigationController {
-
-    func pushViewController(viewController: UIViewController,
-        animated: Bool, completion: Void -> Void) {
-
-            CATransaction.begin()
-            CATransaction.setCompletionBlock(completion)
-            pushViewController(viewController, animated: animated)
-            CATransaction.commit()
-    }
-
-    func popViewController(animated: Bool, completion: Void -> Void) {
-
-            CATransaction.begin()
-            CATransaction.setCompletionBlock(completion)
-            popViewControllerAnimated(animated)
-            CATransaction.commit()
-    }
-    
-}
-
